requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.10.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
autopep8==2.0.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.3
better-exceptions==0.3.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colored==1.3.93
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
DBUtils==3.0.2
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
exceptiongroup==1.1.0
executing==1.2.0
fake-headers==1.0.2
feapder==1.7.9
ffmpeg-python==0.2.0
filelock==3.7.0
Flask==2.1.1
flickrapi==2.4.0
frozenlist==1.3.3
future==0.18.2
geographiclib==2.0
geopy==2.3.0
gunicorn==19.9.0
h11==0.13.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.3.0
incremental==21.3.0
influxdb==5.3.1
ipython==8.10.0
iso8601==1.0.2
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
loguru==0.6.0
lxml==4.8.0
m3u8==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
msgpack==1.0.4
multidict==6.0.2
multitasking==0.0.11
numpy==1.24.2
oauthlib==3.2.2
outcome==1.2.0
packaging==21.0
pandas==1.5.2
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.0
prompt-toolkit==3.0.33
Protego==0.2.1
proxyscrape==0.3.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.14.0
pymongo==4.3.3
PyMySQL==1.0.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
queuelib==1.6.2
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
Scrapy==2.6.1
scrapy-proxy-pool==0.1.9
selenium==4.8.2
service-identity==21.1.0
six==1.15.0
sniffio==1.2.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.1
stack-data==0.6.2
terminal-layout==2.1.2
tldextract==3.3.1
tomli==2.0.1
tqdm==4.63.1
traitlets==5.9.0
trio==0.21.0
trio-websocket==0.10.2
Twisted==22.4.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.13
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
Werkzeug==2.1.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
yfinance==0.1.89
zipp==3.15.0
zope.interface==5.4.0